+++ ddhub_gateway_client/model/paginated_response.py
@@ -87,9 +87,9 @@ class PaginatedResponse(ModelNormal):
         """
         lazy_import()
         return {
-            'count': (float,),  # noqa: E501
-            'limit': (float,),  # noqa: E501
-            'page': (float,),  # noqa: E501
+            'count': (int,),  # noqa: E501
+            'limit': (int,),  # noqa: E501
+            'page': (int,),  # noqa: E501
             'records': ([GetTopicDto],),  # noqa: E501
         }
 
@@ -116,9 +116,9 @@ class PaginatedResponse(ModelNormal):
         """PaginatedResponse - a model defined in OpenAPI
 
         Args:
-            count (float): total number of channels
-            limit (float): limit of channels
-            page (float): page number out of total pages
+            count (int): total number of channels
+            limit (int): limit of channels
+            page (int): page number out of total pages
             records ([GetTopicDto]): Topics records
 
         Keyword Args:
